default new levels to 'none' compression
zlib corruption bug is happening on two different people's computers now, so it's not just a one-off. compression solution is not hugely needed since level data is not the majority of game size, but if someone with a known good PC wants to go around and compress levels or find an alternate solution for bad PCs, then great.